--- packages/store/experimental/index.ts
@@ -1,1 +1,0 @@
-export { NgxsTesting } from './testing/src/symbol';
