--- packages/store/experimental/package.json
@@ -1,10 +1,0 @@
-{
-  "ngPackage": {
-    "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
-    "lib": {
-      "entryFile": "index.ts",
-      "flatModuleFile": "ngxs-store-experimental",
-      "umdModuleIds": {}
-    }
-  }
-}
